Unit again game UI and logic + bug fix
