+++ app/src/main/AndroidManifest.xml
@@ -41,7 +41,7 @@
         <activity android:name=".Login"
             android:exported="false"  />
 
-        <activity android:name=".GameUI"
+        <activity android:name=".GameLogic"
             android:screenOrientation="portrait"
             tools:ignore="DiscouragedApi,LockedOrientationActivity"/>
